Revert "Nn 2203 Tidy up node index "
