+++ backend/controllers/bulkAppointmentsUpload.js
@@ -11,9 +11,7 @@ const bulkAppointmentsUploadFactory = (csvParserService, offenderLoader, logErro
   const index = async (req, res) => {
     const { data } = req.session
 
-    if (!data) {
-      return renderError(req, res)
-    }
+    if (!data) return renderError(req, res)
 
     const appointmentDetails = {
       ...data,
