--- backend/setupBodyParsers.js
@@ -1,19 +1,0 @@
-const express = require('express')
-const bodyParser = require('body-parser')
-const formData = require('express-form-data')
-const os = require('os')
-
-const router = express.Router()
-
-module.exports = () => {
-  router.use(
-    formData.parse({
-      uploadDir: os.tmpdir(),
-      autoClean: true,
-    })
-  )
-  router.use(bodyParser.urlencoded({ extended: false, limit: '5mb', parameterLimit: 1000000 }))
-  router.use(bodyParser.json())
-
-  return router
-}
